setup-environment-internal: stop showing NXP's EULA for TI devices
By default, if not previously accepted the script would prompt to show
NXP's EULA and would then assign ACCEPT_FSL_EULA into one of our local
configs.

This is only needed on NXP's devices, so when we're dealing with TI SoC
(AM62) we simply take that the EULA was previously accepted, so the
script can just skip this whole section.

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>