multiconfig/k3r5: change TMPDIR for k3r5 build
We're facing many dependencies issues with AM62 due to its multiconfig
nature. A fix proposed on the upstream addresses these issues by
changing the TMPDIR used for k3r5 builds, which solves all the
dependency issues we're having on Kirkstone, namely CVE and SBOM
generation.

Since we don't know when upstream will merge this fix, we chose to add
it to our layer, and unblock our CVE and SBOM generation. Once the fix
is merged upstream we can revert this commit.

Upstream fix:
https://patchwork.yoctoproject.org/project/ti/patch/20231112190743.1496752-1-denis@denix.org/

Related-to: TOR-3229

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>