docker-compose: drop docker-compose shell wrapper
"exec docker compose" would not work when docker binary is provided by
podman, because 'docker compose' explicitly depends on docker engine
and must be run as a plugin of it.

Let's change to build a standalone docker-compose by removing
'docker-plugin' pkgconfig, so it could work for both podman and docker
engines.

After this change, both 'docker-compose' and 'docker compose' are
supported for docker and podman engines, when an end user run
docker-compose, it will directly execute docker-compose binary, when
an end user run 'docker compose', docker and podman will locate
'docker-compose' from PATH.

Related-to: TOR-3282

Signed-off-by: Ming Liu <ming.liu@toradex.com>