torizon-conf: revert to using stable tags for containers
This commit is like a soft-revert of 66cce52 done manually because we
weren'table to complete our testing suite of the rc tags with LAVA and thus
we cannot properly release them.

In the meantime, revert to using the stable major 3 tags for the
new torizon release. The impact is minimal because for almost all cases
except for the AM62 platform (that doesn't have 3 major tags), the rc
and non-rc containers are identical.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>