podman: tweak podman.service
The podman system service creates a listening service that answers API
calls for Podman, by default the session is expired after 5 seconds per
run, this sometimes causes failures if the requesting app needs a longer
time to start, like docker-compose, let's pass "-t 0" to podman to avoid
session expires.

Make podman.service an alias of docker.service, we have some systemd
services are depending on docker.service, so we dont have to modify
them to handle the case when podman is the container engine.

Related-to: TOR-3282

Signed-off-by: Ming Liu <ming.liu@toradex.com>