podman: allow torizon user run podman as sudoer without passwd
There are some known shortcomings when running Podman in a rootless
environment:
https://github.com/containers/podman/blob/main/rootless.md

and it's not easy to share images between rootless users and root,
some references:
https://github.com/containers/podman/issues/5608
https://www.redhat.com/sysadmin/image-stores-podman

Although setting 'additionalimagestores' in
/etc/containers/storage.conf allows container images to be shared
among users, but it's still not convenient to use, for instance, if
a root user pulls a image, it has to run
"sudo chmod -R a+rx /var/lib/mycontainers" everytime to avoid
permissions issues. And 'additionalimagestores' only supports
read-only images to be shared among rootless users, an user could not
build a image and share it to other users.

This patch mainly aims to create a docker-like UX for torizon user when
podman is the container engine, by allowing torizon user be able to run
/usr/bin/podman with sudoer without needing input a passwd.

So torizon user can run podman's docker wrapper to access any images
created by root, or build a image to be shared by root, this gives an
end user consistent UX just like he is using docker command.

Related-to: TOR-3284

Signed-off-by: Ming Liu <ming.liu@toradex.com>